+++ DataLayerProjects/PostgresDataLayer/PostGresData.cs
@@ -10,85 +10,11 @@ using System.Linq;
 
 namespace SpoonCMSCore.PostGresData
 {
-    public class PostGresData : ISpoonData
+    public class PostGresData : PostGresDataReader, ISpoonData
     {
-        private DocumentStore _store { get; }
-        private string _connString { get; set; }
-
-        public PostGresData(string connString)
-        {        
-            _connString = connString;
-
-            _store = DocumentStore.For(x =>
-            {
-                x.Connection(_connString);
-                x.Serializer(new JsonNetSerializer { EnumStorage = EnumStorage.AsInteger });
-
-                x.Schema.For<Container>().Duplicate(y => y.Id, configure: y => y.IsUnique = true);
-
-                x.AutoCreateSchemaObjects = AutoCreate.CreateOrUpdate;
-            });
-        }        
-
-        public Container GetContainer(string conName)
+        public PostGresData(string connString): base(connString)
         {
-            try
-            {
-                using (var session = _store.LightweightSession())
-                {
-                    Container container = session.Query<Container>()
-                                            .Where(x => x.Name == conName)
-                                            .FirstOrDefault();
-
-                    return container;
-                }
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        public Container GetContainer(int conId)
-        {
-            try
-            {
-                using (var session = _store.LightweightSession())
-                {
-                    Container container = session.Query<Container>()
-                                            .Where(x => x.Id == conId)
-                                            .FirstOrDefault();
-
-                    return container;
-                }
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        public List<ContainerSkinny> GetAllContainers() //Need to make this lighterweight, this could get heavy
-        {
-            List<ContainerSkinny> retList = new List<ContainerSkinny>();
-            try
-            {
-                using (var session = _store.LightweightSession())
-                {
-                    var containers = session.Query<Container>();
-
-                    foreach (Container con in containers)
-                    {
-                        retList.Add(con.GetSkinny());
-                    }
-                }
-
-                return retList;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
+            
         }
 
         public void AddContainer(Container container)
